+++ internal/install/lock_other.go
@@ -4,6 +4,8 @@ package install
 
 type OperationLock struct{}
 
+func (*OperationLock) heldFor(string) bool { return false }
+
 func AcquireLock(string) (*OperationLock, error) {
 	return nil, installerError("operation locking is unsupported on this host", nil)
 }